# docs/v1/openapi.yml
openapi: "3.0.2"
info:
  title: Sator API
  version: "0.0.1"
servers:
  - url: http://localhost:8080/
    description: Local server
  - url: https://api.dev.sator.io/
    description: Development server
  - url: https://api.stage.sator.io/
    description: Staging server
components:
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ResultSuccess:
      description: "Successful result"
      content:
        application/json:
          schema:
            type: object
            properties:
              result:
                type: boolean
                example: true
...
    DefaultError:
      description: Default error view.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "something went wrong"
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Access token is missing or invalid"
...
paths:
...
  /auth:
    get:
      tags:
        - "Auth"
      summary: Validate access token (JWT).
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ResultSuccess"
        "400":
          $ref: "#/components/responses/DefaultError"
        "403":
          $ref: "#/components/responses/UnauthorizedError"
...
  /profile:
    get:
      tags:
        - "Profile"
      summary: Get user profile data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Returns user profile data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                  username:
                    type: string
                    example: "BigBoy13"
                  first_name:
                    type: string
                    example: "John"
                  last_name:
                    type: string
                    example: "Doe"
                  avatar:
                    type: string
                    example: "https://link@to.avatar"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /balance:
    get:
      tags:
        - "Wallet"
      summary: Get total balance.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Get user total balance across all wallets.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    amount:
                      type: number
                      example: 123.45
                    currency:
                      type: string
                      example: "SAO"
                    sub_amount:
                      type: number
                      example: 50.65
                    sub_currency:
                      type: string
                      example: "Rewards"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /wallets:
    get:
      tags:
        - "Wallet"
      summary: Get wallets list.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Get user wallets list.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                    get_details_url:
                      type: string
                      example: "/wallets/f4f78cac-5db6-4ecc-ad13-5877705f3126"
                    get_transactions_url:
                      type: string
                      example: "/wallets/f4f78cac-5db6-4ecc-ad13-5877705f3126/transactions"
                    order:
                      type: number
                      example: 1
                      description: "Wallet order in list"
        "400":
          $ref: "#/components/responses/DefaultError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /shows:
    get:
      tags:
        - "Shows"
      summary: Get shows list.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: |
            Set needed page number.

            By default, it returns the first page with a set number of items.
          required: false
          schema:
            type: integer
        - name: items_per_page
          in: query
          description: |
            Set needed items per page.

            By default returns 20 items per page.
          required: false
          schema:
            type: integer
        - name: with_nft
          in: query
          description: If with_nft=true it returns list only shows with nft.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Returns shows list with pagination.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                    title:
                      type: string
                      example: "Peaky Blinders"
                    cover:
                      type: string
                      example: "https://m.media-amazon.com/images/M/MV5BZDUwYWJkNTMtYWMzZC00NGE3LWEzYTYtYjAzZjk0M2MzMDQxXkEyXkFqcGdeQXVyNzk4MTAyOTA@._V1_FMjpg_UX1024_.jpg"
                    has_new_episode:
                      type: boolean
                      example: true
                    claps:
                      type: integer
                      example: 234
                    realms_title:
                      type: string
                      example: "Realms"
                    realms_subtitle:
                      type: string
                      example: "Watch on Netflix. Then earn SAO. "
                    watch:
                      type: string
                      example: "https://youtube.com/ERGbevesrf"
                    has_nft:
                      type: boolean
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    post:
      tags:
        - "Shows"
      summary: Used for create new show.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - "title"
                - "cover"
              properties:
                title:
                  type: string
                  example: "title"
                cover:
                  type: string
                  example: "cover"
                has_new_episode:
                  type: boolean
                  example: true
                description:
                  type: string
                  example: "description"
                realms_title:
                  type: string
                  example: "Realms"
                realms_subtitle:
                  type: string
                  example: "Watch on Netflix. Then earn SAO. "
                watch:
                  type: string
                  example: "https://youtube.com/ERGbevesrf"
                categories:
                  type: array
                  items:
                    type: string
                    example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
      responses:
        "200":
          description: Returns shows data.
          content:
            application/json:
              schema:
                type: object
                required:
                  - "title"
                  - "cover"
                properties:
                  id:
                    type: string
                    example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                  title:
                    type: string
                    example: "title"
                  cover:
                    type: string
                    example: "cover"
                  has_new_episode:
                    type: boolean
                    example: true
                  description:
                    type: string
                    example: "description"
                  realms_title:
                    type: string
                    example: "Realms"
                  realms_subtitle:
                    type: string
                    example: "Watch on Netflix. Then earn SAO. "
                  watch:
                    type: string
                    example: "https://youtube.com/ERGbevesrf"
                  categories:
                    type: array
                    items:
                      type: string
                      example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
        "400":
          $ref: "#/components/responses/DefaultError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/DefaultError"
...
  /qrcodes:
    post:
      tags:
        - "QR code"
      summary: used for create new qrcode.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                show_id:
                  type: string
                  example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                episode_id:
                  type: string
                  example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                starts_at:
                  type: string
                  example: "2006-01-02T15:04:05Z"
                  description: "RFC3339"
                expires_at:
                  type: string
                  example: "2006-01-02T15:04:05Z"
                  description: "RFC3339"
                reward_amount:
                  type: number
                  example: 123.56
      responses:
        "200":
          description: Returns qrcode data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                  show_id:
                    type: string
                    example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                  episode_id:
                    type: string
                    example: "f4f78cac-5db6-4ecc-ad13-5877705f3126"
                  starts_at:
                    type: string
                    example: "2006-01-02T15:04:05Z07:00"
                    description: "RFC3339"
                  expires_at:
                    type: string
                    example: "2006-01-02T15:04:05Z07:00"
                    description: "RFC3339"
                  reward_amount:
                    type: number
                    example: 123.56
        "400":
          $ref: "#/components/responses/DefaultError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/DefaultError"
...
  /invitations:
    post:
      tags:
        - "Invitations"
      summary: Sends invitation if person doesn't exist in invitation table.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "qwerty@gmail.com"
      responses:
        "200":
          $ref: "#/components/responses/ResultSuccess"
        "400":
          $ref: "#/components/responses/DefaultError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...